% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{CalculateApplicationRateFilterPaper}
\alias{CalculateApplicationRateFilterPaper}
\title{Calculate Application Rate for filter paper}
\usage{
CalculateApplicationRateFilterPaper(df, rep_value, bottle_number)
}
\arguments{
\item{df}{Dataframe}

\item{rep_value}{Replication value}

\item{bottle_number}{Bottle number}
}
\value{
Dataframe with application rate information
}
\description{
This function calculates the application rate for filter paper.
}
