% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{CalculateStandardCurve}
\alias{CalculateStandardCurve}
\title{Calculate Standard Curve}
\usage{
CalculateStandardCurve(df, rep_value, bottle_number)
}
\arguments{
\item{df}{Dataframe}

\item{rep_value}{Replication value}

\item{bottle_number}{Bottle number}
}
\description{
This function calculates the standard curve.
}
