% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{CalculateVolumeDeposited}
\alias{CalculateVolumeDeposited}
\title{Calculate Application Rate for Petri Dish}
\usage{
CalculateVolumeDeposited(df, rep_value, bottle_number, slope_table)
}
\arguments{
\item{df}{Dataframe}

\item{rep_value}{Replication value}

\item{bottle_number}{Bottle number}
}
\value{
Dataframe with application rate information
}
\description{
This function calculates the application rate for petri dishes.
}
